+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## Current (in progress)
 
 - Remove unused `_total_pages` search property [#2726](https://github.com/opendatateam/udata/pull/2726)
+- Use -followers as default suggest sort on datasets, reuses and orgas [#2727](https://github.com/opendatateam/udata/pull/2727)
 
 ## 4.0.1 (2022-04-11)
 
 - Removed `post_save` signal within `add_resource` and `update_resource` methods. [#2720](https://github.com/opendatateam/udata/pull/2720)
-- Refactor and update changelog with latest udata updates [#2717](https://github.com/opendatateam/udata/pull/2717)
+- Refactor and update documentation with latest udata updates [#2717](https://github.com/opendatateam/udata/pull/2717)
 - Add harvest csv adapter for a catalog of harvesters [#2722](https://github.com/opendatateam/udata/pull/2722)
 
 ## 4.0.0 (2022-03-30)